Move Picker to its own module
